```text Refactor Utils.php to limit maximum back days for school pay sync
This commit modifies the Utils.php file to limit the maximum back days for the school pay sync feature. Previously, the maximum back days were set to 30, but now they have been reduced to 5. This change improves the efficiency of the sync process by reducing the amount of data that needs to be processed. Additionally, it adds an echo statement to display a success message when a transaction is imported.