Refactor SchoolPayTransaction.php and AcademicClass.php models
This commit refactors the SchoolPayTransaction.php and AcademicClass.php models. In SchoolPayTransaction.php, a new relationship method 'account' is added to define the relationship with the Account model. In AcademicClass.php, the 'getNameTextAttribute' method is updated to use the 'short_name' property instead of 'name'. These changes improve the clarity and functionality of the models.